IN-10 refactor: changes names of files for Checkbox
