+++ shared/components/Checkbox/Сheckbox.module.scss
