+++ shared/components/Checkbox/Сheckbox.stories.tsx
@@ -1,6 +1,6 @@
 import { type Meta, type StoryObj } from '@storybook/react'
 import { useState } from 'react'
-import { Checkbox } from './Сheckbox'
+import { Checkbox } from './Checkbox'
 
 const meta = {
   component: Checkbox,
