+++ shared/components/index.ts
@@ -9,7 +9,7 @@ export * from './Textarea'
 export * from './Input'
 export * from './Navbar'
 export * from './SignIn'
-export * from './Сheckbox'
+export * from './Checkbox'
 export * from './Controlled'
 export * from './CreateNewPassword'
-
+export * from './Tabs'
