--- shared/components/Сheckbox/index.ts
@@ -1,1 +1,0 @@
-export * from './Сheckbox'
